requirements.txt
django
djangorestframework
django-rest-swagger
coreapi
